Moves CreatePolicyWizard
